Style + consistency fixes
